requirements.txt
pandas==1.0.1
numpy==1.18.1
SQLAlchemy==1.3.13
psycopg2-binary==2.8.4
